Don't enable `backup-to-temp-directory` by default
Since this setting stores backup files in the volatile /tmp directory, a
user could lose data if their computer crashes. That's not a very
sensible default!

However, some users might still find it useful. This commit removes it
from the set of default settings, but leaves it as an optional,
non-default setting.